requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Machine Learning
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
scipy>=1.7.0

# CLIP (install separately via pip install git+https://github.com/openai/CLIP.git)
ftfy>=6.0.0
regex>=2021.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Data Export
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# Utilities
tqdm>=4.62.0
pathlib>=1.0.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0
